Added mainpage obj and refactored names of files
